Add order template migration and harden password hashing
