+++ backend/app/core/security.py
@@ -8,7 +8,10 @@ from passlib.context import CryptContext
 
 from .config import settings
 
-_pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+_pwd_context = CryptContext(
+    schemes=["bcrypt_sha256", "bcrypt"],
+    deprecated="auto",
+)
 
 
 class TokenType:
