+++ backend/app/db/base.py
@@ -2,6 +2,7 @@ from .models.base import Base
 from .models.broadcast import Broadcast
 from .models.menu import MenuItem, MenuWeek
 from .models.order import Order
+from .models.order_template import OrderTemplate, OrderTemplateWeek
 from .models.order_window import OrderWindow
 from .models.user import User
 
@@ -11,6 +12,8 @@ __all__ = [
     "MenuItem",
     "MenuWeek",
     "Order",
+    "OrderTemplate",
+    "OrderTemplateWeek",
     "OrderWindow",
     "User",
 ]
