+++ backend/pyproject.toml
@@ -24,6 +24,7 @@ dependencies = [
     "httpx>=0.27.0",
     "structlog>=24.1.0",
     "passlib[bcrypt]>=1.7.4",
+    "bcrypt>=3.2,<4.0",
     "pyjwt[crypto]>=2.8.0",
     "email-validator>=2.1.0"
 ]
